feature/21 - Criação do mapper para gerenciamento dos DTOs, implementação e testes com validators

+++ src/main/java/br/com/amorim/supermarket/model/productdata/ProductData.java
@@ -68,7 +68,7 @@ public class ProductData {
 
     @Column(nullable = false)
     @NotNull(message = "Estoque não pode estar vazio")
-    private BigInteger inventory;
+    private BigDecimal inventory;
 
     @ManyToOne
     @JoinColumn(name = "subsection_id")
